shuffle things round
