+++ typescript/infra/scripts/warp-routes/monitor/types.ts
@@ -1,0 +1,12 @@
+export interface xERC20Limit {
+  tokenName: string;
+  mint: number;
+  burn: number;
+  mintMax: number;
+  burnMax: number;
+}
+
+export interface WarpRouteInfo {
+  balance: number;
+  valueUSD?: number;
+}
