+++ typescript/infra/scripts/warp-routes/monitor/utils.ts
@@ -1,0 +1,3 @@
+import { rootLogger } from '@hyperlane-xyz/utils';
+
+export const logger = rootLogger.child({ module: 'warp-balance-monitor' });
